refactor(renderer): split oversized files for #678 PR
CI's per-file 500-line limit flagged three files after the spatial-index
addition:

* `renderer3d/spatial_index.rs` (579 lines) -> `spatial_index/mod.rs`
  (335 lines) plus `spatial_index/tests.rs` (243 lines).
* `renderer3d/tests.rs` (621 lines) -> `tests/mod.rs` (420 lines, the
  pre-existing renderer3d test bundle) plus
  `tests/scene_culling_678.rs` (193 lines, the new scaling/parity
  module).
* `renderer3d/types.rs` (508 lines): extracted `Renderer3DStats` to
  `stats.rs` and re-exported it from `types.rs` so the public path
  `renderer3d::Renderer3DStats` is unchanged.

No behavior change.
