+++ goud_engine/src/libs/graphics/renderer3d/mod.rs
@@ -44,6 +44,7 @@ mod shaders;
 mod shadow;
 mod skinned_mesh;
 mod spatial_index;
+mod stats;
 mod texture;
 mod types;
